requirements.txt
appdirs>=1.4.4
asgiref>=3.3.1
backcall>=0.2.0
crispy-forms-materialize>=0.2
decorator>=4.4.2
distlib>=0.3.1
Django>=3.1.4
django-admin-material>=0.1.0
django-appconf>=1.0.4
django-compat>=1.0.15
django-crispy-forms>=1.10.0
django-material>=1.7.3
django-material-admin>=1.7.16
django-materialize-css>=0.0.1
django-model-utils>=4.1.1
django-pandas>=0.6.2
django-sesame>=2.1
django-statici18n>=2.0.1
django-transmeta-eh>=0.7.6
fef-questionnaire>=6.0
filelock>=3.0.12
ipython>=7.19.0
ipython-genutils>=0.2.0
itsdangerous>=1.1.0
jedi>=0.17.2
numpy>=1.19.4
packaging>=20.7
pandas>=1.1.4
parso<0.8.0
pexpect>=4.8.0
pickleshare>=0.7.5
Pillow>=8.0.1
pluggy>=0.13.1
prompt-toolkit>=3.0.8
ptyprocess>=0.6.0
py>=1.9.0
Pygments>=2.7.2
pyparsing>=2.4.7
python-dateutil>=2.8.1
pytz>=2020.4
PyYAML>=5.3.1
six>=1.15.0
sqlparse>=0.4.1
toml>=0.10.2
tox>=3.20.1
traitlets>=5.0.5
virtualenv>=20.2.1
wcwidth>=0.2.5
